+++ src/utils.h
@@ -1,0 +1,1 @@
+int is_lazy_load(SEXP x);